refund-charges/python/pyproject.toml
[tool.poetry]
name = "interval-starter"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
interval-sdk = "^1.0.0"
python-dotenv = "^0.21.1"


[tool.poetry.group.dev.dependencies]
watchdog = {extras = ["watchmedo"], version = "^2.2.1"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
